AddPost screen hotfixes with ImageModel

+++ app/src/main/java/com/example/glazovnetadminapp/domain/models/ImageModel.kt
@@ -2,6 +2,6 @@ package com.example.glazovnetadminapp.domain.models
 
 data class ImageModel(
     val imageUrl: String,
-    val imageWidth: Float,
-    val imageHeight: Float
+    val imageWidth: Int,
+    val imageHeight: Int
 )
