+++ app/src/main/java/com/example/glazovnetadminapp/entity/ImageModelDto.kt
@@ -6,7 +6,7 @@ data class ImageModelDto(
     @field:Json(name = "imageUrl")
     val imageUrl: String,
     @field:Json(name = "imageWidth")
-    val imageWidth: Float,
+    val imageWidth: Int,
     @field:Json(name = "imageHeight")
-    val imageHeight: Float
+    val imageHeight: Int
 )
