+++ app/src/main/java/com/example/glazovnetadminapp/presentation/posts/postsList/PostCard.kt
@@ -99,7 +99,7 @@ fun PostCard(
                     maxLines = descriptionMaxLines
                 )
                 post.image?.let { image ->
-                    val imageAspectRatio = ( image.imageWidth / image.imageHeight )
+                    val imageAspectRatio = image.imageWidth.toFloat() / image.imageHeight.toFloat()
                     Spacer(modifier = Modifier.height(10.dp))
                     AsyncImage(
                         model = image.imageUrl,
